+++ R/pcr_peak_calls.R
@@ -20,10 +20,11 @@ pcr_peak_calls <- function(peaks, assay) {
         size_delta_bp = .data$size_bp - .data$expected_size_bp,
         within_window = .data$size_bp >= .data$lower_size_bp & .data$size_bp <= .data$upper_size_bp,
         above_min_concentration = .data$concentration >= .data$min_concentration,
-        evidence_zone = dplyr::if_else(
-            .data$above_min_concentration,
-            "above_analytical",
-            "below_analytical"
+        above_confirm_concentration = .data$concentration >= .data$confirm_concentration,
+        evidence_zone = dplyr::case_when(
+            .data$above_confirm_concentration ~ "above_confirmatory",
+            .data$above_min_concentration ~ "analytical_to_confirmatory",
+            TRUE ~ "below_analytical"
         ),
         matched = .data$within_window & .data$above_min_concentration
     )